+++ Dimitri/Game/Game/src/Models/Updatables/LightBeam.h
@@ -17,7 +17,7 @@ namespace Game {
 		*/
 		class LightBeam : public Models::Updatable {
 		public:
-			LightBeam(int x, int y, int z, int height, int width, Enums::DirectionEnum state, Graphics::Models::Center center);
+			LightBeam(float x, float y, float z, float height, float width, Enums::DirectionEnum state, Graphics::Models::Center center);
 
 			/**
 			* \brief Initializes Textures
